requirements.txt
certifi==2018.11.29
chardet==3.0.4
Click==7.0
coreapi==2.3.3
coreapi-cli==1.0.9
coreschema==0.0.4
Django==2.1.7
django-crispy-forms==1.7.2
django-vote==2.1.7
djangorestframework==3.9.1
djangorestframework-simplejwt==4.0.0
httpie==1.0.2
idna==2.8
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.1.1
mysqlclient==1.4.2.post1
Pillow==5.4.1
pkg-resources==0.0.0
psycopg2==2.7.7
psycopg2-binary==2.7.7
Pygments==2.3.1
PyJWT==1.7.1
pytz==2018.9
requests==2.21.0
uritemplate==3.0.0
urllib3==1.24.1
votes==1.0.1
gunicorn==19.7.1


